requirements.txt
yfinance==0.1.54
ipykernel==5.3.4
scikit-learn==0.23.1
papermill==2.1.2
apache-airflow-providers-papermill==1.0.2
gitdb==4.0.5
gitpython==3.1.7
apache-airflow-providers-http==2.0.0
pytest==6.0.1